update drop shadows in status badges
we inverted the text color on many of the badges we got from sheilds.io
in order to achieve a better contrast ratio.  This change inverts the
accompanying text drop shadows on those badges as well so that light
text has a dark shadow, and dark text has a light shadow (which is what
shields.io does by default).  This corrects a visual fuzziness on some
badges that occurs when dark text has a dark shadow.